+++ src/main/clojure/com/github/clojure_lsp/intellij/extension/general_settings.clj
@@ -7,8 +7,11 @@
    [clojure.walk :as walk]
    [com.github.clojure-lsp.intellij.client :as lsp-client]
    [com.github.clojure-lsp.intellij.db :as db]
-   [seesaw.core :as see]
-   [seesaw.mig :as mig])
+   [com.github.clojure-lsp.intellij.server :as server]
+   [seesaw.color :as s.color]
+   [seesaw.core :as s]
+   [seesaw.font :as s.font]
+   [seesaw.mig :as s.mig])
   (:import
    [com.github.clojure_lsp.intellij.extension SettingsState]
    [com.intellij.ui IdeBorderFactory]
@@ -23,26 +26,41 @@
 (def ^:private server-not-started-message "Server not started")
 
 (defn ^:private build-component [{:keys [server-version log-path] :as server-info}]
-  (mig/mig-panel
-   :items [[(mig/mig-panel :border (IdeBorderFactory/createTitledBorder "Troubleshooting")
-                           :items [[(see/label "Server log path") ""]
-                                   [(see/text :id :server-log
-                                              :columns 30
-                                              :editable? false
-                                              :text (or log-path server-not-started-message)) "wrap"]
-                                   [(see/label "Server version") ""]
-                                   [(see/text :id :server-log
-                                              :columns 30
-                                              :editable? false
-                                              :text (or server-version server-not-started-message)) "wrap"]
-                                   [(see/label "Server trace level (Requires LSP restart)") ""]
-                                   [(see/combobox :id :trace-level :model ["off" "messages" "verbose"]) "wrap"]
-                                   [(see/button :text "Copy server info to clipboard"
-                                                :listen [:action (fn [e]
-                                                                   (.setContents (.getSystemClipboard (Toolkit/getDefaultToolkit))
-                                                                                 (StringSelection. (with-out-str (pprint/pprint server-info)))
-                                                                                 nil)
-                                                                   (see/alert e "Server info copied to clipboard"))]) ""]]) "span"]]))
+  (let [server-running? (boolean server-info)]
+    (s.mig/mig-panel
+     :items (->> [(when-not server-running?
+                    [(s/label :text "Warning: Clojure LSP is not running or not started yet"
+                              :foreground (s.color/color 243 156 18)) "wrap"])
+                  [(s.mig/mig-panel :border (IdeBorderFactory/createTitledBorder "Troubleshooting")
+                                    :items [[(s/label "Server log path") ""]
+                                            [(s/text :id :server-log
+                                                     :columns 30
+                                                     :editable? false
+                                                     :enabled? server-running?
+                                                     :text (or log-path server-not-started-message)) "wrap"]
+                                            [(s/label "Server version") ""]
+                                            [(s/text :id :server-log
+                                                     :columns 30
+                                                     :editable? false
+                                                     :enabled? server-running?
+                                                     :text (or server-version server-not-started-message)) "wrap"]
+                                            [(s/label "Server trace level") ""]
+                                            [(s/combobox :id :trace-level :model ["off" "messages" "verbose"]) "split 2"]
+                                            [(s/label :text "requires LSP restart"
+                                                      :font (s.font/font  :size 14)
+                                                      :foreground (s.color/color 110 110 110)) "wrap"]
+                                            [(s/button :text "Copy server info to clipboard"
+                                                       :listen [:action (fn [e]
+                                                                          (.setContents (.getSystemClipboard (Toolkit/getDefaultToolkit))
+                                                                                        (StringSelection. (with-out-str (pprint/pprint server-info)))
+                                                                                        nil)
+                                                                          (s/alert e "Server info copied to clipboard"))]) "wrap"]
+                                            [(s/button :text "Restart LSP server"
+                                                       :listen [:action (fn [_]
+                                                                          (server/shutdown!)
+                                                                          (server/spawn-server! (:project @db/db*)))]) "wrap"]]) "span"]
+                  [(s/label "Tip: clojure-lsp is built-in this plugin, external installation is not needed.") "shrink"]]
+                 (remove nil?)))))
 
 (defn -createComponent [_]
   (let [server-info (some-> (lsp-client/connected-client)
@@ -54,15 +72,15 @@
     component))
 
 (defn -getPreferredFocusedComponent [_]
-  (see/select @component* [:#trace-level]))
+  (s/select @component* [:#trace-level]))
 
 (defn -isModified [_]
   (let [settings-state (SettingsState/get)
-        trace-level-combo-box ^JComboBox (see/select @component* [:#trace-level])]
+        trace-level-combo-box ^JComboBox (s/select @component* [:#trace-level])]
     (not= (.getSelectedItem trace-level-combo-box) (.getTraceLevel settings-state))))
 
 (defn -reset [_]
-  (let [trace-level-combo-box ^JComboBox (see/select @component* [:#trace-level])]
+  (let [trace-level-combo-box ^JComboBox (s/select @component* [:#trace-level])]
     (.setSelectedItem trace-level-combo-box (-> @db/db* :settings :trace-level))))
 
 (defn -disposeUIResources [_]
@@ -70,7 +88,7 @@
 
 (defn -apply [_]
   (let [settings-state (SettingsState/get)
-        trace-level-combo-box ^JComboBox (see/select @component* [:#trace-level])]
+        trace-level-combo-box ^JComboBox (s/select @component* [:#trace-level])]
     (db/set-trace-level-setting! settings-state (.getSelectedItem trace-level-combo-box))))
 
 (defn -cancel [_])